+++ src/main/java/com/gitlab/cdagaming/craftpresence/utils/CommandUtils.java
@@ -61,6 +61,7 @@ public class CommandUtils {
      */
     public static void reloadData(final boolean forceUpdateRPC) {
         ModUtils.TRANSLATOR.onTick();
+        CraftPresence.SYSTEM.setRefreshRate(CraftPresence.CONFIG.refreshRate);
         CraftPresence.SYSTEM.onTick();
         CraftPresence.instance.addScheduledTask(() -> CraftPresence.KEYBINDINGS.onTick());
         CraftPresence.GUIS.onTick();