+++ Changes.md
@@ -22,6 +22,7 @@ available [here](https://gitlab.com/CDAGaming/CraftPresence/-/compare/release%2F
 
 ### Fixes
 
+* Fixed an issue where refresh rates under 2 seconds could brick the `SystemUtils` module, causing the RPC to break
 * Backend: Added additional String<->UUID conversion utils to ensure anything involving UUIDs follows Mojang's format
 * Backend: Rewritten `UrlUtils#openUrl` to support Linux + Modern Versions of Windows and Java
 * Backend: Fixed `DiscordUtils#updateTimestamp` not properly working